Refactored the grpc Android app
