+++ gradle/libs.versions.toml
@@ -26,11 +26,13 @@ protobufGradlePlugin = "0.8.18"
 android = { module = "com.google.android:android", version.ref = "android" }
 androidx-annotations = { module = "androidx.annotation:annotation", version = "1.1.0" }
 androidx-appcompat = { module = "androidx.appcompat:appcompat", version = "1.0.2" }
+androidx-compose = { module = "androidx.activity:activity-compose", version = "1.4.0" }
 androidx-constraintLayout = { module = "androidx.constraintlayout:constraintlayout", version = "1.1.3" }
 androidx-ktx = { module = "androidx.core:core-ktx", version = "1.0.2" }
 animalSniffer-annotations = { module = "org.codehaus.mojo:animal-sniffer-annotations", version.ref = "animalSniffer" }
 animalSniffer-gradle = { module = "ru.vyarus:gradle-animalsniffer-plugin", version.ref = "animalSnifferGradle" }
 assertj = { module = "org.assertj:assertj-core", version.ref = "assertj" }
+contour = { module = "app.cash.contour:contour", version = "1.1.0" }
 dokka-core = { module = "org.jetbrains.dokka:dokka-core", version.ref = "dokka" }
 dokka-gradlePlugin = { module = "org.jetbrains.dokka:dokka-gradle-plugin", version.ref = "dokka" }
 grpc-netty = { module = "io.grpc:grpc-netty", version.ref = "grpc" }
