--- samples/wire-grpc-sample/client/build.gradle.kts
@@ -1,29 +1,0 @@
-plugins {
-  id("com.android.application")
-}
-
-android {
-  compileSdkVersion(30)
-  defaultConfig {
-    applicationId = "com.squareup.wire.whiteboard"
-    minSdkVersion(28)
-    targetSdkVersion(30)
-  }
-  compileOptions {
-    sourceCompatibility = JavaVersion.VERSION_1_8
-    targetCompatibility = JavaVersion.VERSION_1_8
-  }
-  lintOptions {
-    isCheckReleaseBuilds = false // CI runs full lint.
-  }
-}
-
-dependencies {
-  implementation(projects.samples.wireGrpcSample.protos)
-  implementation(libs.kotlin.coroutines.android)
-  implementation(libs.androidx.appcompat)
-  implementation(libs.androidx.ktx)
-  implementation(libs.wire.grpcClient)
-  implementation(libs.androidx.appcompat)
-  implementation(libs.androidx.constraintLayout)
-}
