+++ samples/wire-grpc-sample/client/src/main/AndroidManifest.xml
@@ -1,19 +1,22 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.squareup.wire.whiteboard">
+  xmlns:tools="http://schemas.android.com/tools"
+  package="com.squareup.wire.whiteboard">
 
   <uses-permission android:name="android.permission.INTERNET"/>
   <application
-      android:allowBackup="true"
-      android:icon="@mipmap/ic_launcher"
-      android:label="@string/app_name"
-      android:roundIcon="@mipmap/ic_launcher_round"
-      android:supportsRtl="true"
-      android:theme="@style/AppTheme">
-    <activity android:name="com.squareup.wire.whiteboard.MainActivity">
+    android:allowBackup="true"
+    android:icon="@mipmap/ic_launcher"
+    android:label="@string/app_name"
+    android:roundIcon="@mipmap/ic_launcher_round"
+    android:supportsRtl="true"
+    tools:targetApi="31">
+    <activity
+      android:name=".MainActivity"
+      android:exported="true"
+      android:label="@string/app_name">
       <intent-filter>
         <action android:name="android.intent.action.MAIN"/>
-
         <category android:name="android.intent.category.LAUNCHER"/>
       </intent-filter>
     </activity>
