--- samples/wire-grpc-sample/client/src/main/res/drawable/border_gb.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-  <stroke
-      android:color="#2b2b2b"
-      android:width="2dp" />
-</shape>
