--- samples/wire-grpc-sample/client/src/main/res/layout/main_activity.xml
@@ -1,29 +1,0 @@
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    >
-  <view
-      class="com.squareup.wire.whiteboard.MainActivity$WhiteboardView"
-      android:id="@+id/whiteboard"
-      android:layout_width="match_parent"
-      android:layout_height="0dp"
-      android:background="@drawable/border_gb"
-      app:layout_constraintBottom_toTopOf="@+id/clear"
-      app:layout_constraintTop_toTopOf="parent"
-      />
-  <androidx.appcompat.widget.AppCompatButton
-      android:id="@+id/clear"
-      android:layout_width="wrap_content"
-      android:layout_height="wrap_content"
-      android:layout_margin="24dp"
-      android:text="Clear !"
-      android:textSize="30sp"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toBottomOf="@+id/whiteboard"
-      tools:ignore="HardcodedText"
-      />
-</androidx.constraintlayout.widget.ConstraintLayout>
